requirements.txt
asgiref==3.5.2
Django==4.1.1
mysql-connector-python==8.0.30
protobuf==3.20.1
sqlparse==0.4.2
